Develop merges StephanieQ

+++ Algoritmos/files.cpp
@@ -2,13 +2,13 @@
 #include <string>
 
 using namespace std;
-void readFile(string filename, float values[],int size){
+void readFile(string filename, double values[],int size){
     ifstream archivo(filename);
 	char linea[128];
     int index = 0;
     while(index !=size){
         archivo.getline(linea, sizeof(linea));
-        values[index] = stof(linea);
+        values[index] = stod(linea);
         index++;
     }
     archivo.close();
